pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "First try at building an AI agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
